fix(billing): move billing data to separate file
